Organize JS files into folders
