+++ source/__tests__/BreakPrompt.test.js
@@ -1,5 +1,5 @@
 import { jest, test } from '@jest/globals';
-import BreakPrompt from '../js/BreakPrompt.js';
+import BreakPrompt from '../js/components/BreakPrompt.js';
 
 describe('BreakPrompt testing', () => {
   let jestFunc; // jest mock function
