+++ source/__tests__/StartButtons.test.js
@@ -1,6 +1,6 @@
 import { describe, expect, test } from '@jest/globals';
-import StartButtons from '../js/StartButtons.js';
-import TaskList from '../js/TaskList.js';
+import StartButtons from '../js/components/StartButtons.js';
+import TaskList from '../js/classes/TaskList.js';
 
 describe('StartButtons tests', () => {
   test('Simple Constructor', () => {
