+++ source/__tests__/StartContainer.test.js
@@ -1,5 +1,5 @@
 import { describe, expect } from '@jest/globals';
-import StartContainer from '../js/StartContainer.js';
+import StartContainer from '../js/components/StartContainer.js';
 
 describe('StartContainer tests', () => {
   beforeEach(() => {
