+++ source/__tests__/Timer.test.js
@@ -1,7 +1,7 @@
 // to learn jest.timers -> https://jestjs.io/docs/en/timer-mocks
 
 import { jest } from '@jest/globals';
-import Timer from '../js/Timer.js';
+import Timer from '../js/classes/Timer.js';
 
 // CONSTRUCTOR TESTS
