+++ source/__tests__/TimerUI.test.js
@@ -1,5 +1,5 @@
 import { expect, jest, test } from '@jest/globals';
-import TimerUI from '../js/TimerUI.js';
+import TimerUI from '../js/components/TimerUI.js';
 
 // Initialize the DOM with a TimerUI element.
 beforeEach(() => {
