+++ source/__tests__/UsernameInput.test.js
@@ -1,5 +1,5 @@
 import { expect } from '@jest/globals';
-import UsernameInput from '../js/UsernameInput.js';
+import UsernameInput from '../js/components/UsernameInput.js';
 
 // Initialize the DOM with a UsernameInput element.
 beforeEach(() => {
