+++ source/__tests__/WelcomeMessage.test.js
@@ -1,5 +1,5 @@
 import { beforeEach, jest, test } from '@jest/globals';
-import WelcomeMessage from '../js/WelcomeMessage.js';
+import WelcomeMessage from '../js/components/WelcomeMessage.js';
 
 beforeEach(() => {
   document.body.innerHTML = '<h1> Something </h1>';
