+++ source/app.html
@@ -455,8 +455,8 @@
             </label>
         </div>
     </template>
-    <script src="js/TomatoSlider.js" type="module"></script>
-    <script src="js/app.js" type="module"></script>
+    <script src="js/components/TomatoSlider.js" type="module"></script>
+    <script src="js/scripts/app.js" type="module"></script>
     <noscript>Please enable Javascript!</noscript>
 </body>
