+++ source/done.html
@@ -35,7 +35,7 @@
             </div>
         </div>
     </div>
-    <script src="js/done.js" type="module"></script>
+    <script src="js/scripts/done.js" type="module"></script>
     <noscript>Please enable Javascript!</noscript>
 </body>
