+++ source/index.html
@@ -237,9 +237,9 @@
             </div>
         </div>
     </footer>
-    <script src="js/StartContainer.js" type="module"></script>
-    <script src="js/CurrentYear.js" type="module"></script>
-    <script src="js/LogsButton.js" type="module"></script>
+    <script src="js/components/StartContainer.js" type="module"></script>
+    <script src="js/components/CurrentYear.js" type="module"></script>
+    <script src="js/components/LogsButton.js" type="module"></script>
     <noscript>Please enable Javascript!</noscript>
 </body>
 </html>
