+++ source/js/classes/Timer.js
