+++ source/js/components/WelcomeMessage.js
